feat: implement Branch 13 - Performance Optimization with per-CPU stats, RCU, and slab cache
Implement comprehensive performance optimization system for MUTEX kernel proxy:

New Features:
- Per-CPU statistics for lock-free performance counters
- RCU-protected connection lookups for lock-free reads
- Slab cache memory pool for fast connection allocation
- jhash-based hash optimization for better distribution
- Zero-copy packet handling with skb_clone
- Packet batching support for bulk operations
- Comprehensive performance monitoring and profiling

Files Added:
- src/module/mutex_perf_opt.h (267 lines) - Performance optimization APIs
- src/module/mutex_perf_opt.c (1,089 lines) - Full implementation
- docs/BRANCH_13_SUMMARY.md (394 lines) - Complete documentation
- docs/PERFORMANCE_OPT_QUICK_REF.md - Developer quick reference

Integration:
- Modified mutex_conn_track.c to use performance optimizations
- Modified mutex_proxy_core.c to initialize perf subsystem
- Updated Makefile to include mutex_perf_opt.o

Compatibility Fixes (Kernel 6.8 API):
- Fixed mutex_process_filter.c: cgroup_path_ns → cgroup_path
- Fixed mutex_process_filter.c: task_session_nr → task_session_vnr (3x)
- Fixed mutex_process_filter.c: get_mm_exe_file → direct mm->exe_file with RCU
- Fixed mutex_process_filter.c: find_task_by_vpid → pid_task(find_vpid()) (6x)
- Fixed mutex_perf_opt.c: removed floating point operations (3x)
- Fixed mutex_protocol_detect.h: separated internal/userspace stats structs
- Fixed mutex_protocol_detect.c: atomic64_t vs u64 type mismatches

Build Status: ✅ Successfully compiles to mutex_proxy.ko (3.3MB)

Implements: Branch 13 from BRANCH_PLAN.md
Fixes: All compilation errors across branches 11, 12, and 13

+++ src/module/mutex_protocol_detect.c
@@ -943,25 +943,25 @@ void protocol_detect_get_stats(struct protocol_detect_context *ctx,
 		return;
 
 	for (i = 0; i < PROTO_MAX; i++) {
-		stats->proto_detected[i] = atomic64_read(&ctx->stats.proto_detected[i]);
-		stats->proto_errors[i] = atomic64_read(&ctx->stats.proto_errors[i]);
+		stats->proto_detected[i] = (u64)atomic64_read(&ctx->stats.proto_detected[i]);
+		stats->proto_errors[i] = (u64)atomic64_read(&ctx->stats.proto_errors[i]);
 	}
 
-	stats->method_port_hits = atomic64_read(&ctx->stats.method_port_hits);
-	stats->method_pattern_hits = atomic64_read(&ctx->stats.method_pattern_hits);
-	stats->method_heuristic_hits = atomic64_read(&ctx->stats.method_heuristic_hits);
-	stats->method_dpi_hits = atomic64_read(&ctx->stats.method_dpi_hits);
-	stats->method_sni_hits = atomic64_read(&ctx->stats.method_sni_hits);
-	stats->method_handshake_hits = atomic64_read(&ctx->stats.method_handshake_hits);
+	stats->method_port_hits = (u64)atomic64_read(&ctx->stats.method_port_hits);
+	stats->method_pattern_hits = (u64)atomic64_read(&ctx->stats.method_pattern_hits);
+	stats->method_heuristic_hits = (u64)atomic64_read(&ctx->stats.method_heuristic_hits);
+	stats->method_dpi_hits = (u64)atomic64_read(&ctx->stats.method_dpi_hits);
+	stats->method_sni_hits = (u64)atomic64_read(&ctx->stats.method_sni_hits);
+	stats->method_handshake_hits = (u64)atomic64_read(&ctx->stats.method_handshake_hits);
 
-	stats->routed_proxy = atomic64_read(&ctx->stats.routed_proxy);
-	stats->routed_direct = atomic64_read(&ctx->stats.routed_direct);
-	stats->routed_blocked = atomic64_read(&ctx->stats.routed_blocked);
+	stats->routed_proxy = (u64)atomic64_read(&ctx->stats.routed_proxy);
+	stats->routed_direct = (u64)atomic64_read(&ctx->stats.routed_direct);
+	stats->routed_blocked = (u64)atomic64_read(&ctx->stats.routed_blocked);
 
-	stats->total_packets = atomic64_read(&ctx->stats.total_packets);
-	stats->total_inspections = atomic64_read(&ctx->stats.total_inspections);
-	stats->cache_hits = atomic64_read(&ctx->stats.cache_hits);
-	stats->cache_misses = atomic64_read(&ctx->stats.cache_misses);
+	stats->total_packets = (u64)atomic64_read(&ctx->stats.total_packets);
+	stats->total_inspections = (u64)atomic64_read(&ctx->stats.total_inspections);
+	stats->cache_hits = (u64)atomic64_read(&ctx->stats.cache_hits);
+	stats->cache_misses = (u64)atomic64_read(&ctx->stats.cache_misses);
 }
 
 /**
